# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: E-Credentialing CVO Platform — Public REST API v1
  version: "1.8.0"
  summary: Read-only public REST API for the E-Credentialing CVO platform.
  description: |
    The v1 surface is intentionally read-only. Every request requires
    a Bearer API key with the appropriate scope; every request is
    audited; PHI fields (SSN, DOB, DEA, personal address, notes) are
    NEVER returned regardless of scope.

    For a no-auth playground using synthetic data, see
    `/api/sandbox/v1/*` (documented at `/sandbox`).

    Rate limiting (since v1.2.0): every successful response carries
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and
    `X-RateLimit-Reset` headers (Unix-seconds). When you exceed the
    per-key budget the API returns a `429` with the `RateLimitProblem`
    body and a `Retry-After` header.

    Request correlation (since v1.3.0): every request and response —
    success or error — carries an `X-Request-Id` header. Send your
    own (`^[A-Za-z0-9_\-]{8,128}$`) and we'll honour it; otherwise
    we generate one server-side. Paste it into a support ticket and
    we can find your call in O(1). See `docs/api/versioning.md`
    for the full contract and deprecation policy.

    Key introspection (since v1.4.0): `GET /api/v1/me` returns the
    current API key's name, granted scopes, lifecycle timestamps,
    and current rate-limit budget. Pairs with `/health` to make
    "is my key configured correctly?" a one-call answer.

...

...

...

...

...
  contact:
...
servers:
...
security:
  - BearerApiKey: []
...
paths:
...
  /api/v1/providers:
    get:
      tags: [providers]
      summary: List providers
      description: |
        Returns a paginated list of providers. Supports filtering by
        `status` and `npi`. PHI is never included.
      operationId: listProviders
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IfNoneMatchHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/ProviderStatus" }
        - in: query
          name: npi
          description: Exact-match NPI filter.
          schema: { type: string, pattern: "^[0-9]{10}$" }
      responses:
        "200":
          description: Page of providers.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            Link: { $ref: "#/components/headers/Link" }
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
            Deprecation: { $ref: "#/components/headers/Deprecation" }
            Sunset: { $ref: "#/components/headers/Sunset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvidersPage" }
        "304": { $ref: "#/components/responses/NotModified" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/providers/{id}:
    get:
      tags: [providers]
      summary: Get one provider (full envelope, no PHI)
      operationId: getProvider
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IfNoneMatchHeader"
        - in: path
          name: id
          required: true
          schema: { type: string, format: cuid }
      responses:
        "200":
          description: Provider envelope.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
            Deprecation: { $ref: "#/components/headers/Deprecation" }
            Sunset: { $ref: "#/components/headers/Sunset" }
            Link: { $ref: "#/components/headers/Link" }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/ProviderDetail" }
        "304": { $ref: "#/components/responses/NotModified" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/v1/sanctions:
    get:
      tags: [sanctions]
      summary: List sanctions checks
      operationId: listSanctions
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/IfNoneMatchHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: providerId
          schema: { type: string, format: cuid }
        - in: query
          name: result
          schema: { $ref: "#/components/schemas/SanctionsResult" }
      responses:
        "200":
          description: Page of sanctions checks.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
            Link: { $ref: "#/components/headers/Link" }
            Deprecation: { $ref: "#/components/headers/Deprecation" }
            Sunset: { $ref: "#/components/headers/Sunset" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SanctionsPage" }
        "304": { $ref: "#/components/responses/NotModified" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: |
        Issue keys via `/admin/api-keys`. Each key carries one or more
        scopes (see `x-scopes` below). Keys are stored hashed
        (SHA-256) in `api_keys`; raw value is shown once at creation.
      x-scopes:
        providers:read: Read provider summary + detail records.
        providers:cv: Generate the curriculum vitae PDF for a provider.
        sanctions:read: Read OIG/SAM sanction match records.
        enrollments:read: Read payor enrollment status records.
        documents:read: Read provider document metadata (no PHI).
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      description: |
        Optional caller-supplied correlation id. If supplied and it
        matches `^[A-Za-z0-9_\-]{8,128}$` the server honours it and
        echoes it back on the response. Otherwise the server
        generates one. Pair this with your client-side log lines for
        end-to-end tracing. Available since v1.3.0.
      schema: { type: string, pattern: "^[A-Za-z0-9_\\-]{8,128}$" }
    IfNoneMatchHeader:
      in: header
      name: If-None-Match
      required: false
      description: |
        Conditional GET — supply the `ETag` value the server returned
        on a previous response and, if nothing has changed, the
        server replies `304 Not Modified` with an empty body. Either
        the literal `*` (wildcard, always matches a current
        representation) or one or more quoted ETag tokens
        (`W/"<hex>"` or `"<hex>"`), comma-separated. Weak
        comparison per RFC 9110 §13.1.2 — a strong-form token
        (`"abc"`) matches the corresponding weak-form ETag
        (`W/"abc"`). Available since v1.6.0.
      schema: { type: string }
      example: 'W/"deadbeefcafebabe1234567890abcdef12345678"'
    Page:
      in: query
      name: page
      description: 1-based page index. Default 1.
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      in: query
      name: limit
      description: Items per page (max 100). Default 25.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  responses:
    Unauthorized:
      description: Missing or invalid API key. Body is RFC 9457 Problem (since v1.8.0).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
        Deprecation: { $ref: "#/components/headers/Deprecation" }
        Sunset: { $ref: "#/components/headers/Sunset" }
        Link: { $ref: "#/components/headers/Link" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: API key valid but lacks the required scope. Body is RFC 9457 Problem (since v1.8.0).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
        Deprecation: { $ref: "#/components/headers/Deprecation" }
        Sunset: { $ref: "#/components/headers/Sunset" }
        Link: { $ref: "#/components/headers/Link" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found. Body is RFC 9457 Problem (since v1.8.0).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
        Deprecation: { $ref: "#/components/headers/Deprecation" }
        Sunset: { $ref: "#/components/headers/Sunset" }
        Link: { $ref: "#/components/headers/Link" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotModified:
      description: |
        Conditional GET cache hit (since v1.6.0). The caller sent
        `If-None-Match` with an ETag that matches the current
        representation. The response body is empty (zero bytes per
        RFC 9110 §15.4.5); the `ETag`, `X-Request-Id`, and
        rate-limit headers are still present, and (since v1.7.0)
        any `Deprecation` / `Sunset` / `Link` advisory headers
        applicable to the operation also fire on the 304 — a
        client polling on `If-None-Match` must still find out
        about deprecation.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        Deprecation:
          $ref: "#/components/headers/Deprecation"
        Sunset:
          $ref: "#/components/headers/Sunset"
        Link:
          $ref: "#/components/headers/Link"
    RateLimited:
      description: |
        Too many requests for this key. The body is a
        `RateLimitProblem`. The `Retry-After` header (seconds) tells
        you when the next request is allowed; the standard
        `X-RateLimit-*` headers (also present on every successful
        response since v1.2.0) describe your budget.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        Deprecation:
          $ref: "#/components/headers/Deprecation"
        Sunset:
          $ref: "#/components/headers/Sunset"
        Link:
          $ref: "#/components/headers/Link"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/RateLimitProblem" }
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitProblem" }
  headers:
    RequestId:
      description: |
        Correlation id for this request/response pair. If the inbound
        request supplied a valid `X-Request-Id` we honour it;
        otherwise we generate one server-side as `req_<hex>`.
        Present on every response (2xx and non-2xx). Format:
        `^[A-Za-z0-9_\-]{8,128}$`. Pass it to support to look up the
        full audit + log trail in O(1). Available since v1.3.0.
      schema: { type: string, pattern: "^[A-Za-z0-9_\\-]{8,128}$", example: "req_8f3c2a91d4b6e0f5" }
    RateLimitLimit:
      description: |
        Maximum requests allowed in the current fixed window for the
        authenticated API key. Stable across the window. Present on
        every successful 2xx response and on the 429 rejection.
      schema: { type: integer, minimum: 1, example: 120 }
    RateLimitRemaining:
      description: |
        Requests still available in the current window (>= 0).
        Decrements on every accepted request. On a 429 this is `0`.
      schema: { type: integer, minimum: 0, example: 117 }
    RateLimitReset:
      description: |
        Unix-seconds (UTC) timestamp when the current window resets
        and `X-RateLimit-Remaining` returns to `X-RateLimit-Limit`.
      schema: { type: integer, minimum: 0, example: 1739887200 }
    RetryAfter:
      description: |
        Seconds the client should wait before retrying. Always >= 1.
        Present only on `429` responses.
      schema: { type: integer, minimum: 1, example: 7 }
    Link:
      description: |
        RFC 8288 web links, comma-separated. Carries up to three
        independent classes of entries; **only the entries that
        apply to a given response are emitted**. Use the SDK's
        `parseLinkHeader(headers.get("Link"))` helper to turn it
        into a `{ rel: url }` map.

          1. **Pagination** (since v1.5.0). Present on every
             paginated list response when the result set has at
             least one page. Always emits `first` and `last`;
             emits `prev` when `page > 1` and `next` when
             `page < totalPages`.
          2. **Deprecation pointers** (since v1.7.0). When the
             current operation is on a deprecation path, emits
             `rel="deprecation"` and `rel="sunset"` entries
             pointing at the upgrade guide, plus an optional
             `rel="successor-version"` entry pointing at the
             replacement endpoint when one exists.

        The pagination + deprecation entries are independent: a
        deprecated paginated endpoint emits BOTH classes of links
        in the same `Link` header.
      schema:
        type: string
        example: |-
          <https://api.example.com/api/v1/providers?page=1&limit=25>; rel="first", <https://api.example.com/api/v1/providers?page=2&limit=25>; rel="next", <https://api.example.com/api/v1/providers?page=4&limit=25>; rel="last", <https://app.example.com/changelog#legacy-thing>; rel="deprecation"
    ETag:
      description: |
        Weak entity-tag for the response body, derived from a stable
        SHA-1 hash of the canonicalized cacheable subset of the
        payload. Always in `W/"<40-hex>"` form. Echo this back as
        `If-None-Match: "<etag>"` on the next poll to receive a
        `304 Not Modified` when nothing has changed. The hash is
        computed AFTER excluding per-request fields (timestamps,
        rate-limit snapshots, last-used markers) so the validator
        is stable across cosmetic refreshes. Available since v1.6.0.
      schema:
        type: string
        pattern: '^W/"[a-f0-9]{40}"$'
        example: 'W/"deadbeefcafebabe1234567890abcdef12345678"'
    Deprecation:
      description: |
        RFC 9745 advisory header. **Present only when this operation
        has entered its deprecation window** — its absence means the
        operation is on the supported path. Value is the
        structured-fields integer form: an `@`-prefixed Unix-seconds
        timestamp recording when deprecation took effect (e.g.
        `@1796083200` is `2026-12-01T00:00:00Z`).

        When `Deprecation` is present, `Sunset` is also present, and
        the `Link` header carries `rel="deprecation"` (and where
        applicable `rel="sunset"`, `rel="successor-version"`)
        entries pointing at the upgrade guide and replacement
        endpoint. Available since v1.7.0.
      schema:
        type: string
        pattern: '^@[0-9]+$'
        example: "@1796083200"
    Sunset:
      description: |
        RFC 8594 advisory header. **Present only when this operation
        has entered its deprecation window** — pairs with the
        `Deprecation` header. Value is an HTTP-date (IMF-fixdate per
        RFC 9110 §5.6.7) marking the wall-clock at which the
        operation will start returning `410 Gone`. Always at least
        180 days in the future at first publication; the policy
        lives in `docs/api/versioning.md` §4. Available since
        v1.7.0.
      schema:
        type: string
        example: "Sun, 11 Nov 2030 23:59:59 GMT"
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, error]
      description: |
        Canonical v1 error body. Since v1.8.0 every non-2xx response
        body is an RFC 9457 Problem object that ALSO carries the
        legacy `{ error: { code, message, ...extras } }` envelope as
        an extension member, so existing SDKs continue to read
        `body.error.code` unchanged. New clients SHOULD read the
        top-level RFC 9457 fields (`type`, `title`, `status`,
        `detail`, `instance`).

        The Content-Type is `application/problem+json` per RFC 9457
        §3, except when the caller's `Accept` header excludes it —
        in which case the same body is returned with
        `Content-Type: application/json`.
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI per error code. Per RFC 9457 §3.1.1 this is
            "the primary identifier for the problem type". Changing
            this URI for an existing code is a breaking change.
          example: "https://essen-credentialing.example/errors/not-found"
        title:
          type: string
          description: |
            Short human-readable summary. Per RFC 9457 §3.1.3 this
            SHOULD NOT change between occurrences of the problem.
          example: "Resource not found"
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: |
            HTTP status code, mirrored into the body for consumers
            who only see the body (e.g. behind proxies that strip
            status). MUST equal the response status.
          example: 404
        detail:
          type: string
          description: |
            Human-readable explanation specific to this occurrence.
            Mirrors `error.message`.
          example: "Provider not found"
        instance:
          type: string
          description: |
            URI reference identifying the specific occurrence —
            typically the request path that triggered the error.
          example: "/api/v1/providers/abc"
        error:
          type: object
          required: [code, message]
          description: |
            Legacy envelope, preserved verbatim for backward
            compatibility. Removing this is a breaking change that
            requires a new major version.
          properties:
            code:
              type: string
              description: |
                Stable machine-readable identifier for the failure
                mode. Snake_case. Part of the contract — never renamed
                without bumping `info.version` and following the
                deprecation policy in `docs/api/versioning.md`.
              example: "not_found"
            message:
              type: string
              description: Human-readable explanation, may change between releases.
              example: "Provider not found"
            required:
              type: string
              description: |
                Present on `insufficient_scope` errors — names the
                scope the API key would have needed. Optional on
                other errors.
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    ProviderStatus:
      type: string
      description: Lifecycle stage of the provider record.
      enum:
        - INVITED
        - APPLICATION_IN_PROGRESS
        - VERIFICATION_IN_PROGRESS
        - COMMITTEE_REVIEW
        - APPROVED
        - DENIED
        - TERMINATED
        - SUSPENDED
    SanctionsResult:
      type: string
      enum: [CLEAR, MATCH, REVIEW_REQUIRED]
    EnrollmentStatus:
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - SUBMITTED
        - APPROVED
        - DENIED
        - TERMINATED
...
    ProviderSummary:
      type: object
      required: [id, legalFirstName, legalLastName, status, createdAt]
      properties:
        id: { type: string, format: cuid }
        legalFirstName: { type: string }
        legalLastName: { type: string }
        npi: { type: string, nullable: true, pattern: "^[0-9]{10}$" }
        status: { $ref: "#/components/schemas/ProviderStatus" }
        providerType: { $ref: "#/components/schemas/ProviderType" }
        approvedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ProviderDetail:
      allOf:
        - $ref: "#/components/schemas/ProviderSummary"
        - type: object
          properties:
            legalMiddleName: { type: string, nullable: true }
            caqhId: { type: string, nullable: true }
            updatedAt: { type: string, format: date-time }
            profile:
              type: object
              nullable: true
              properties:
                specialtyPrimary: { type: string, nullable: true }
                facilityAssignment: { type: string, nullable: true }
                jobTitle: { type: string, nullable: true }
            licenses:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: cuid }
                  state: { type: string }
                  licenseType: { type: string }
                  status: { type: string }
                  expirationDate: { type: string, format: date, nullable: true }
            enrollments:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: cuid }
                  payerName: { type: string }
                  status: { $ref: "#/components/schemas/EnrollmentStatus" }
                  effectiveDate: { type: string, format: date, nullable: true }
            expirables:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: cuid }
                  expirableType: { type: string }
                  status: { type: string }
                  expirationDate: { type: string, format: date, nullable: true }
    ProvidersPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ProviderSummary" }
        pagination: { $ref: "#/components/schemas/Pagination" }
    SanctionsCheck:
      type: object
      required: [id, source, result, runDate]
      properties:
        id: { type: string, format: cuid }
        source: { type: string, example: "OIG_LEIE" }
        result: { $ref: "#/components/schemas/SanctionsResult" }
        runDate: { type: string, format: date-time }
        triggeredBy: { type: string, nullable: true }
        provider:
          type: object
          properties:
            id: { type: string, format: cuid }
            legalFirstName: { type: string }
            legalLastName: { type: string }
            npi: { type: string, nullable: true }
    SanctionsPage:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/SanctionsCheck" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...